refactor(cli): extract runtime output writer
